pyproject.toml
[project]
name = "qq"
version = "0.0.2"
description = "qq – local context server + CLI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [{name = "core tools"}]
dependencies = [
  "typer>=0.12",
  "pyyaml>=6.0",
  "pydantic>=2.7",
  "httpx>=0.27",
  "fastapi>=0.111",
  "uvicorn>=0.30",
  "qdrant-client>=1.9",
  "rich>=13.7",
  "sentence-transformers>=3.0",
  "numpy>=1.26",
  "tqdm>=4.66",
  "rank-bm25>=0.2.2",
  "openai>=1.40.0",
  "google-generativeai>=0.7.2",
]

[project.scripts]
qq = "qq.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qq"]
